+++ src/test/ts/browser/LoadTinyTest.ts
@@ -3,9 +3,10 @@ import { UnitTest } from '@ephox/bedrock-client';
 import { Arr, Strings, Global } from '@ephox/katamari';
 import { SelectorFilter, Attribute, SugarElement, Remove } from '@ephox/sugar';
 
-import { ScriptLoader } from '../../../main/ts/ScriptLoader';
+import { ScriptLoader } from '../../../main/ts/ScriptLoader2';
 import { cRemove, cRender } from '../alien/Loader';
 
+const apiKey = 'qagffr3pkuv17a8on1afax661irst1hbr4e6tbv888sz91jc';
 UnitTest.asynctest('LoadTinyTest', (success, failure) => {
   const cDeleteTinymce = Chain.op(() => {
     ScriptLoader.reinitialize();
@@ -71,5 +72,50 @@ UnitTest.asynctest('LoadTinyTest', (success, failure) => {
       cRemove,
       cDeleteTinymce,
     ]),
+    Log.chainsAsStep('Should be able to load TinyMCE in hybrid', '', [
+      cRender({ tinymceScriptSrc: [
+        '/project/node_modules/tinymce-6/tinymce.min.js',
+        `https://cdn.tiny.cloud/1/${apiKey}/tinymce/6/cloud-plugins.min.js?tinydrive=6`
+      ], plugins: [ 'tinydrive' ] }),
+      cAssertTinymceVersion('6'),
+      Chain.op(() => {
+        Assertions.assertEq(
+          'TinyMCE should have been loaded locally',
+          '/project/node_modules/tinymce-6',
+          Global.tinymce.baseURI.path
+        );
+      }),
+      Chain.op(() => {
+        Assertions.assertEq(
+          'The tinydrive plugin should have defaults for the cloud',
+          `https://cdn.tiny.cloud/1/${apiKey}/tinymce-plugins/tinydrive/6/plugin.min.js`,
+          Global.tinymce.defaultOptions?.custom_plugin_urls?.tinydrive
+        );
+      }),
+      cRemove,
+      cDeleteTinymce,
+
+      cRender({ tinymceScriptSrc: [
+        '/project/node_modules/tinymce-5/tinymce.min.js',
+        `https://cdn.tiny.cloud/1/${apiKey}/tinymce/5/cloud-plugins.min.js?tinydrive=5`
+      ], plugins: [ 'tinydrive' ] }),
+      cAssertTinymceVersion('5'),
+      Chain.op(() => {
+        Assertions.assertEq(
+          'TinyMCE should have been loaded locally',
+          '/project/node_modules/tinymce-5',
+          Global.tinymce.baseURI.path
+        );
+      }),
+      Chain.op(() => {
+        Assertions.assertEq(
+          'The tinydrive plugin should have defaults for the cloud',
+          `https://cdn.tiny.cloud/1/${apiKey}/tinymce-plugins/tinydrive/5/plugin.min.js`,
+          Global.tinymce.defaultSettings?.custom_plugin_urls?.tinydrive
+        );
+      }),
+      cRemove,
+      cDeleteTinymce,
+    ])
   ], success, failure);
 });